pyproject.toml
[tool.poetry]
name = "selenium-2captcha-solver"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "selenium_2captcha_solver"}]

[tool.poetry.dependencies]
python = ">3.8"
selenium = "^4.10.0"
2captcha-python = "^1.2.0"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.2"
isort = "^5.12.0"
pytest = "^7.3.2"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
